Add files via upload
In this project, I built a model to predict Titanic passenger survival. I handled missing values by filling Age with the median, imputing Embarked with the most common value, and dropping the Cabin column. I created a FamilySize feature and used label encoding for Sex and Embarked. A Random Forest classifier was trained, evaluated using AUC, and made predictions on the test set. The results were submitted in the required format for evaluation.